+++ eslint.config.js
@@ -7,7 +7,7 @@ import tseslint from 'typescript-eslint'
 import reactHooks from 'eslint-plugin-react-hooks'
 
 export default tseslint.config(
-  { ignores: ['dist', 'node_modules', 'public'] },
+  { ignores: ['dist', 'node_modules', 'public', 'storybook-static'] },
   js.configs.recommended,
   ...tseslint.configs.recommended,
   {